Set up Docker-based Python-Mapnik command line utility
The initial commit includes the setup of the command line utility using Docker. Python-Mapnik is encapsulated in the Docker image, and is further provided with the required scripts for building, running and validating. Additional scripts for install and run are included. The commit also adds a README, .gitignore, and docker-compose file for orchestration purposes.